Revert OwnablePermissions changes
