+++ udata/core/dataservices/permissions.py
@@ -1,4 +1,6 @@
-from udata.core.owned import OwnablePermission
+from udata.core.dataset.permissions import (
+    OwnablePermission
+)
 
 class DataserviceEditPermission(OwnablePermission):
     '''Permissions to edit a Dataservice'''
