+++ udata/core/topic/permissions.py
@@ -1,4 +1,4 @@
-from udata.core.owned import OwnablePermission
+from udata.core.dataset.permissions import OwnablePermission
 
 
 class TopicEditPermission(OwnablePermission):
